Split source code into separate files and add ES module distribution.
