+++ src/index.ts
@@ -1,0 +1,7 @@
+import { visitString } from "./visitString";
+
+export * from "./StringVisitor";
+export * from "./StringVisitee";
+export * from "./visitString";
+
+export default visitString;
